requirements.txt
requests>=2.31.0
pillow>=10.0.0
schedule>=1.2.0
numpy>=1.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0
zeroconf>=0.71.0